+++ glpi_python_client/clients/api/knowledgebase/_category.py
@@ -68,7 +68,7 @@ class KBCategoryMixin(TransportMixin):
 
         Raises
         ------
-        ValueError
+        GlpiStatusError
             If the GLPI server returns a non-success HTTP status.
         """